ci: use .java-version and prepare to use VM with installed tools in BK

+++ .ci/release.sh
@@ -21,5 +21,10 @@ echo $PATH
 java -version
 
 set +x
-echo "--- Deploy the release :package:"
-./mvnw -V -s .ci/settings.xml -Pgpg clean deploy -DskipTests --batch-mode | tee release.txt
+if [[ "$dry_run" == "false" ]] ; then
+  echo "--- Deploy the release :package:"
+  ./mvnw -V -s .ci/settings.xml -Pgpg clean deploy -DskipTests --batch-mode | tee release.txt
+else
+  echo "--- Deploy the release :package: (dry-run)"
+  ./mvnw -V -s .ci/settings.xml -Pgpg clean install -DskipTests --batch-mode | tee release.txt
+fi
